+++ copy_with_extension/lib/copy_with_extension.dart
@@ -5,5 +5,9 @@ library copy_with_extension;
 /// Annotation used to indicate that the `copyWith` extension should be generated.
 class CopyWith {
   /// Creates a new [CopyWith] instance.
-  const CopyWith();
+  /// Set `generateCopyWithNull` to `true` for generating an extra `copyWithNull` function that allows you to nullify the properties.
+  const CopyWith({this.generateCopyWithNull = false})
+      : assert(generateCopyWithNull is bool);
+
+  final bool generateCopyWithNull;
 }