+++ copy_with_extension_test/lib/basic_class.g.dart
@@ -1,28 +1,3 @@
-const String pkgName = "copy_with_extension_gen";
-
-const String annotationsBase = r'''
-class CopyWith {
-  const CopyWith({this.generateCopyWithNull = false})
-      : assert(generateCopyWithNull is bool);
-
-  final bool generateCopyWithNull;
-}
-''';
-
-const String correctInput = r'''
-import 'package:copy_with_extension/copy_with_extension.dart';
-
-part 'basic_class.g.dart';
-
-@CopyWith(generateCopyWithNull: true)
-class BasicClass {
-  final String id;
-
-  BasicClass({this.id});
-}
-''';
-
-const String correctResult = r'''
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 part of 'basic_class.dart';
@@ -50,4 +25,3 @@ extension BasicClassCopyWithExtension on BasicClass {
     );
   }
 }
-''';